Enabled big instances loading, minor fixes
- New loading routine:
  - "-f1|-f2" argument before filename indicating file loading routine
  - Changed all tour lengths variables from int to int64_t
    due to overflow in big instances
- Fixed a typo in "usage" string

+++ TSP/Algorithms.h
@@ -16,10 +16,10 @@ public:
 	static void neighbour_insert(int* tour, int i1, int i2);
 
 private:
-	static int compute_tour_length(Instance* instance, int* tour);
-	static void display_tour_length(int tour_length, int optimal_tour_length);
-	static void display_tour(int tour_length, int optimal_tour_length, int* tour, int size);
-	static void permute(Instance* instance, int* a, int l, int r, int* best_tour_length, int* best_tour);
+	static int64_t compute_tour_length(Instance* instance, int* tour);
+	static void display_tour_length(int64_t tour_length, int64_t optimal_tour_length);
+	static void display_tour(int64_t tour_length, int64_t optimal_tour_length, int* tour, int size);
+	static void permute(Instance* instance, int* a, int l, int r, int64_t* best_tour_length, int* best_tour);
 	static int* generate_random_tour(int size);
 	static void random_shuffle_array(int* a, int size);
 	static void calculate_fitness(Individual* individual, Instance* instance);
